requirements.txt
boto3
pinecone-client
langchain
langchain-pinecone
langchain-openai
langchain-community
fastapi
uvicorn
psycopg2-binary

